+++ build.sh
@@ -8,7 +8,7 @@ then
     exit $exit_code
   fi
 
-  .paket/paket.exe install -v
+  .paket/paket.exe restore
   exit_code=$?
   if [ $exit_code -ne 0 ]; then
     exit $exit_code
@@ -24,7 +24,7 @@ else
     exit $exit_code
   fi
 
-  mono .paket/paket.exe install -v
+  mono .paket/paket.exe restore
   exit_code=$?
   if [ $exit_code -ne 0 ]; then
     exit $exit_code