+++ src/Common/AssemblyInfo.fs
@@ -4,10 +4,19 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("FSharp.Formatting")>]
 [<assembly: AssemblyProductAttribute("FSharp.Formatting")>]
 [<assembly: AssemblyDescriptionAttribute("A package of libraries for building great F# documentation, samples and blogs")>]
+<<<<<<< HEAD
 [<assembly: AssemblyVersionAttribute("2.4.25")>]
 [<assembly: AssemblyFileVersionAttribute("2.4.25")>]
+=======
+[<assembly: AssemblyVersionAttribute("2.4.35")>]
+[<assembly: AssemblyFileVersionAttribute("2.4.35")>]
+>>>>>>> ee5a1d473420797c01ccf48aee18a158efc82b4c
 [<assembly: AssemblyCopyrightAttribute("Apache 2.0 License")>]
 do ()
 
 module internal AssemblyVersionInformation =
+<<<<<<< HEAD
     let [<Literal>] Version = "2.4.25"
+=======
+    let [<Literal>] Version = "2.4.35"
+>>>>>>> ee5a1d473420797c01ccf48aee18a158efc82b4c