+++ src/FSharp.Literate/Formatting.fs
@@ -1,9 +1,9 @@
 ﻿namespace FSharp.Literate
 
 open System.IO
+open System.Collections.Concurrent
 open System.Globalization
 open FSharp.Literate
-open RazorEngine.Templating
 open FSharp.CodeFormat
 open FSharp.Markdown
 
@@ -68,14 +68,18 @@ module Templating =
           html.Replace("{" + key + id + "}", value)) input
         result 
 
+  /// Very basic and simple caching mechanism for RazorRender instances.
+  let private razorCache = new ConcurrentDictionary<string[] * string, RazorRender>()
+
   /// Depending on the template file, use either Razor engine
   /// or simple Html engine with {replacements} to format the document
   let private generateFile contentTag parameters templateOpt output layoutRoots =
     match templateOpt with
     | Some (file:string) when file.EndsWith("cshtml", true, CultureInfo.InvariantCulture) -> 
-        let razor = RazorRender(layoutRoots, [])
+        let key = Array.ofSeq layoutRoots, file
+        let razor = razorCache.GetOrAdd(key, fun (layoutRoots, file) -> RazorRender(layoutRoots, [], Path.GetFileNameWithoutExtension file))
         let props = [ "Properties", dict parameters ]
-        let generated = razor.ProcessFile(file, props)
+        let generated = razor.ProcessFile(props)
         File.WriteAllText(output, generated)      
     | _ ->
         let templateOpt = templateOpt |> Option.map File.ReadAllText