+++ tests/FSharp.MetadataFormat.Tests/Tests.fs
@@ -162,3 +162,41 @@ let ``MetadataFormat process XML comments in two sample F# assemblies``() =
   files.["fslib-nested.html"] |> should contain "Somewhat nested module"
   files.["fslib-nested-nestedtype.html"] |> should contain "Very nested member"
   files.["fslib-nested-submodule.html"] |> should contain "Very nested field"
+
+[<Test>]
+let ``MetadataFormat handles c# dlls`` () =
+  let library = root @@ "files" @@ "CSharpFormat.dll"
+  let output = getOutputDir()
+  MetadataFormat.Generate
+    ( library, output, layoutRoots, info, libDirs = [root @@ "../../lib"])
+  let files = Directory.GetFiles(output)
+  
+  let optIndex = files |> Seq.tryFind (fun s -> s.EndsWith "index.html")
+  optIndex.IsSome |> shouldEqual true
+
+  #if INTERACTIVE
+  System.Diagnostics.Process.Start(output)
+  #endif
+
+[<Test>]
+let ``MetadataFormat processes C# types and includes xml comments in docs`` () =
+    let library = root @@ "files" @@ "CSharpFormat.dll"
+    let output = getOutputDir()
+    MetadataFormat.Generate
+        ( library, output, layoutRoots, info, libDirs = [root @@ "../../lib"])
+    let fileNames = Directory.GetFiles(output)
+    let files = dict [ for f in fileNames -> Path.GetFileName(f), File.ReadAllText(f) ]
+    files.["index.html"] |> should contain "CLikeFormat"
+    files.["index.html"] |> should contain "Provides a base class for formatting languages similar to C."
+
+[<Test>]
+let ``MetadataFormat processes C# properties on types and includes xml comments in docs`` () =
+    let library = root @@ "files" @@ "CSharpFormat.dll"
+    let output = getOutputDir()
+    MetadataFormat.Generate
+        ( library, output, layoutRoots, info, libDirs = [root @@ "../../lib"])
+    let fileNames = Directory.GetFiles(output)
+    let files = dict [ for f in fileNames -> Path.GetFileName(f), File.ReadAllText(f) ] 
+    
+    files.["manoli-utils-csharpformat-clikeformat.html"] |> should contain "CommentRegEx"
+    files.["manoli-utils-csharpformat-clikeformat.html"] |> should contain "Regular expression string to match single line and multi-line"